requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
openpyxl>=3.0.0
numpy>=1.21.0
db-dtypes>=1.0.0
